+++ libs/vre/shared/app-representations/src/lib/resource-fetcher.service.ts
@@ -51,6 +51,7 @@ export class ResourceFetcherService {
   }
 
   onDestroy() {
+    this._store.dispatch(new SetCurrentResourceAction(null));
     this._subscription?.unsubscribe();
   }