+++ libs/vre/shared/app-resource-page/src/lib/compound/compound.service.ts
@@ -5,15 +5,23 @@ import { IncomingService } from '@dasch-swiss/vre/shared/app-common-to-move';
 import { DspApiConnectionToken } from '@dasch-swiss/vre/shared/app-config';
 import { NotificationService } from '@dasch-swiss/vre/shared/app-notification';
 import { RegionService } from '@dasch-swiss/vre/shared/app-representations';
+import { Subject } from 'rxjs';
 
 @Injectable()
 export class CompoundService {
-  compoundPosition!: DspCompoundPosition;
+  compoundPosition?: DspCompoundPosition;
   incomingResource: DspResource | undefined;
-  private _resource!: DspResource;
+  resource!: DspResource;
+
+  onOpenNotLoadedIncomingResourcePage$: Subject<void> = new Subject<void>();
 
   get exists() {
     return this.compoundPosition !== undefined;
+  }
+
+  reset() {
+    this.compoundPosition = undefined;
+    this.incomingResource = undefined;
   }
 
   constructor(
@@ -27,18 +35,32 @@ export class CompoundService {
 
   onInit(_compound: DspCompoundPosition, resource: DspResource) {
     this.compoundPosition = _compound;
-    this._resource = resource;
+    this.resource = resource;
     this.openPage(_compound.page);
   }
 
   openPage(page: number) {
+    if (!this.compoundPosition) {
+      return;
+    }
+
     const offset = Math.ceil(page / 25) - 1;
     const position = Math.floor(page - offset * 25 - 1);
 
     // get incoming still image representations, if the offset changed
-    if (offset === this.compoundPosition.offset && this._resource.incomingRepresentations.length > 0) {
+    if (
+      offset === this.compoundPosition.offset &&
+      this.resource.incomingRepresentations.length > 0 &&
+      this.resource.incomingRepresentations[position]
+    ) {
       // get incoming resource, if the offset is the same but page changed
-      this._loadIncomingResource(this._resource.incomingRepresentations[position].id);
+      this._loadIncomingResource(this.resource.incomingRepresentations[position].id);
+    } else if (
+      offset === this.compoundPosition.offset &&
+      this.resource.incomingRepresentations.length > 0 &&
+      !this.resource.incomingRepresentations[position]
+    ) {
+      this.onOpenNotLoadedIncomingResourcePage$.next();
     } else {
       this.compoundPosition.offset = offset;
       this._loadIncomingResourcesPage(offset);
@@ -55,15 +77,15 @@ export class CompoundService {
     }
 
     this._incomingService
-      .getStillImageRepresentationsForCompoundResource(this._resource.res.id, offset)
+      .getStillImageRepresentationsForCompoundResource(this.resource.res.id, offset)
       .subscribe(res => {
         const incomingImageRepresentations = res as ReadResourceSequence;
 
         if (incomingImageRepresentations.resources.length === 0) {
           return;
         }
-        this._resource.incomingRepresentations = incomingImageRepresentations.resources;
-        this._loadIncomingResource(this._resource.incomingRepresentations[this.compoundPosition.position].id);
+        this.resource.incomingRepresentations = incomingImageRepresentations.resources;
+        this._loadIncomingResource(this.resource.incomingRepresentations[this.compoundPosition.position].id);
       });
   }